requirements.txt
construct==2.5.2
cryptography==1.3.1
Django==1.9.6
django-crispy-forms==1.6.0
django-compressor==1.6
django-grappelli==2.8.1
django-haystack==2.5.dev0
django-markdown-deux==1.0.5
django-mptt==0.8.3
django-redis==4.4.3
django-redis-cache==1.7.0
django-uuidfield==0.5.0
enum34==1.1.4
future==0.15.2
beautifulsoup==3.2.1
twilio==6.0.0
httplib2==0.9.2
idna==2.1
ipaddress==1.0.16
lxml==3.6.0
Markdown==2.6.6
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
nltk==3.2.2
oauth2==1.9.0.post1
oauthlib==1.0.3
pefile==2016.3.28
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==16.0.0
python-dateutil==2.5.2
python-openid==2.2.5
python-ptrace==0.9
python-social-auth==0.2.16
python-twitter==2.2
pytz==2016.6.1
requests==2.10.0
requests-oauthlib==0.6.1
redis==2.10.5
six==1.10.0
sqlparse==0.1.19
Whoosh==2.7.4
textile==2.3.16
tomd==0.1.3
